+++ test/qclient.cc
@@ -198,17 +198,17 @@ TEST(QClient, T3) {
 
     RedisRequest incoming;
     for(size_t i = 0; i < 10; i++) {
-      RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+      RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
       ASSERT_EQ(incoming, make_req("PING", std::to_string(i)));
       link.Send(SSTR("+" << i << "\r\n"));
     }
 
-    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
     ASSERT_EQ(incoming, req1);
     link.Send("+OK\r\n");
     ASSERT_REPLY(fut1, "OK");
 
-    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
     ASSERT_EQ(incoming, req2);
     link.Send("+ZZZ\r\n");
     ASSERT_REPLY(fut2, "ZZZ");
@@ -239,16 +239,16 @@ TEST(QClient, AuthHandshake) {
     std::future<redisReplyPtr> fut2 = tunnel.execute(req2);
 
     RedisRequest incoming;
-    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
     ASSERT_EQ(incoming, make_req("AUTH", "hunter2"));
     link.Send("+OK\r\n");
 
-    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
     ASSERT_EQ(incoming, req1);
     link.Send("+OK\r\n");
     ASSERT_REPLY(fut1, "OK");
 
-    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming), 1);
+    RETRY_ASSERT_EQ_SPIN(parser.fetch(incoming, true), 1);
     ASSERT_EQ(incoming, req2);
     link.Send("+ZZZ\r\n");
     ASSERT_REPLY(fut2, "ZZZ");